Moved remaining pages into the 'pages' folder
